% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-km.R
\name{geom_km}
\alias{geom_km}
\title{Kaplan-Meier plot}
\usage{
geom_km(
  mapping = NULL,
  data = NULL,
  stat = "KM",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  na.rm = TRUE,
  ...
)
}
\description{
Kaplan-Meier plot
}
