% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-risktable.R
\name{geom_risktable}
\alias{geom_risktable}
\title{Survival table plot}
\usage{
geom_risktable(
  mapping = NULL,
  data = NULL,
  stat = "RiskTable",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  table.type = "r",
  ...
)
}
\description{
Generates a survival table plot.
}
