% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-risktable.R
\docType{data}
\name{GeomRiskTable}
\alias{GeomRiskTable}
\title{Risk and Event Table}
\format{
An object of class \code{GeomRiskTable} (inherits from \code{GeomText}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomRiskTable
}
\description{
Risk and Event Table
}
\keyword{datasets}
