% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtte-package.R
\docType{package}
\name{ggtte-package}
\alias{ggtte}
\alias{ggtte-package}
\title{ggtte: Implementation of Grammar of Graphics for Time to Event Data}
\description{
This package is an extension of `ggplot2` with specific tools for plotting time-to-event and survival data.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
